src/scraper/youtube.py
import aiohttp
import json
import random
import os
from dotenv import load_dotenv
from .base import BaseScraper

# .env dosyasını yükle
load_dotenv()


class YouTubeScraper(BaseScraper):
    async def scrape(self, keywords, limit=10):
        await self.init_session()
        results = []

        # YouTube API için kimlik bilgileri
        api_key = os.getenv("YOUTUBE_API_KEY")

        # Mock veriler için
        for keyword in keywords:
            # Örnek YouTube verileri
            videos = []
            for i in range(limit):
                vid_id = f"video_{i}_{keyword.replace(' ', '_')}"
                videos.append({
                    "id": vid_id,
                    "title": f"YouTube: {keyword} hakkında video {i}",
                    "view_count": random.randint(1000, 1000000),
                    "like_count": random.randint(100, 50000),
                    "comment_count": random.randint(10, 5000),
                    "channel": f"Channel{i}",
                    "published_at": f"2023-01-{str(i + 1).zfill(2)}T12:00:00Z",
                    "url": f"https://youtube.com/watch?v={vid_id}"
                })

            results.append({
                "keyword": keyword,
                "source": "youtube",
                "data": videos
            })

        # Not: Gerçek YouTube API implementasyonu burada eklenecek

        return results