requirements.txt
typer>=0.7.0
rich>=12.0.0
aiohttp>=3.8.1
aiosqlite>=0.17.0
pandas>=1.4.0
numpy>=1.21.0
pyyaml>=6.0
pytrends>=4.8.0
python-dotenv>=1.0.0